Fix open-archive test and make long-name insertion test more reliable
Open-archive test now uses the current archive format (was hardcoded as
GRP) and the long-name insertion test now checks that the archive doesn't
get corrupted after an insertion failure due to user error (like filename
too long.)

+++ tests/tests.cpp
@@ -33,7 +33,7 @@ void default_sample::printNice(boost::test_tools::predicate_result& res, const s
 	for (int i = 0; i < s.length(); i++) {
 		if (s[i] < 32) {
 			res.message() << "\\x" << std::setfill('0') << std::setw(2)
-				<< std::hex << (int)s[i];
+				<< std::hex << (int)((uint8_t)s[i]);
 		} else {
 			res.message() << s[i];
 		}
